requirements.txt
africastalking==1.1.9
amqp==2.5.2
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.2.3
astroid==2.3.2
Babel==2.7.0
beautifulsoup4==4.6.0
billiard==3.6.1.0
Brlapi==0.6.6
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
command-not-found==0.3
configobj==5.0.6
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
cupshelpers==1.0
defer==1.0.6
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-appconf==1.0.3
django-cors-headers==3.2.1
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.5
django-imagekit==4.0.2
django-markdown-deux==1.0.5
django-minio-storage==0.2.2
django-nine==0.2.3
django-phonenumber-field==3.0.1
django-phonenumbers==1.0.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
future==0.17.1
httplib2==0.9.2
idna==2.8
importlib-metadata==1.4.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
kombu==4.6.7
language-selector==0.1
lazy-object-proxy==1.4.2
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
markdown2==2.3.8
MarkupSafe==1.1.1
mccabe==0.6.1
minio==5.0.1
more-itertools==8.1.0
nemo-emblems==4.0.2
netifaces==0.10.4
oauthlib==3.1.0
onboard==1.4.1
PAM==0.4.2
pexpect==4.2.1
phonenumbers==8.11.1
pilkit==2.0
Pillow==6.1.0
protobuf==3.0.0
psutil==5.4.2
psycopg2-binary==2.8.4
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.4.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.8.1
python-debian==0.1.32
python-decouple==3.1
python-xapp==1.4.0
python-xlib==0.20
python3-openid==3.1.0
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
schema==0.7.1
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.14.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
twilio==6.35.2
typed-ast==1.4.0
ubuntu-drivers-common==0.0.0
ufw==0.36
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
virtualenv==16.7.6
wrapt==1.11.2
xkit==0.0.0
zipp==1.0.0
